refactor(graphql-test-utils): simplify loading schemas
