+++ cypress/support/commands.js
@@ -1,7 +1,6 @@
 import '@percy/cypress';
 import '@testing-library/cypress/add-commands';
-import 'cypress-graphql-mock';
-import '../../graphql-test-utils/cypress-commands';
+import '../../graphql-test-utils/cypress/add-commands';
 
 function isLocalhost() {
   const url = new URL(Cypress.config('baseUrl'));
