+++ graphql-test-utils/cypress/task.js
@@ -1,0 +1,8 @@
+const loadSchema = require('../load-schema');
+
+module.exports = {
+  // This will be available as `cy.task('getGraphQLSchema')`
+  getGraphQLSchema(targetName) {
+    return loadSchema(targetName);
+  },
+};
